+++ connectors/hive1-connector/pom.xml
@@ -67,7 +67,7 @@
         <dependency>
             <groupId>org.apache.hadoop</groupId>
             <artifactId>hadoop-common</artifactId>
-            <version>2.10.1</version>
+            <version>2.7.1</version>
         </dependency>
         <dependency>
             <groupId>org.apache.hadoop</groupId>
