+++ connectors/hive1-connector/src/main/java/io/tapdata/connector/hive1/config/Hive1Config.java
@@ -7,8 +7,18 @@ import java.io.Serializable;
 public class Hive1Config extends CommonDbConfig implements Serializable {
 
 
+    private String hiveConnType;
+
     public Hive1Config() {
         setDbType("hive2");
         setJdbcDriver("org.apache.hive.jdbc.HiveDriver");
+    }
+
+    public String getHiveConnType() {
+        return hiveConnType;
+    }
+
+    public void setHiveConnType(String hiveConnType) {
+        this.hiveConnType = hiveConnType;
     }
 }
