+++ connectors/hive1-connector/src/test/java/io/tapdata/connector/hive1/Demo.java
@@ -1,0 +1,14 @@
+package io.tapdata.connector.hive1;
+
+import org.junit.jupiter.api.Test;
+
+import java.util.Date;
+
+public class Demo {
+
+    @Test
+    public void test1() {
+        Date date = new Date(1661489898L);
+        System.out.println(date);
+    }
+}
